Move ProductSpace and chunker to utils
